Select recursers by matching topic
At the moment, each recurser specifies the days on which they would like
to have a meeting. Pairs of recursers are randomly selected such that
they all want to have a meeting on that day.

The goal here is to add a constraint such that the *intent* of the
meeting can be specified. We can match a pair of recursers together that
would like to discuss a certain topic, for example. So in addition to
having the constraint that the recursers are available to talk on a
given day, we want to constrain the pairing to want to talk about
specific topic.

What:
1. Add topic to database (UNTESTED)
2. Add N^2 algorithm for selecting pair of recursers with matching topic

TODO:
1. Unable to generate list of topics for cleaning user input (see
`parse_cmd.go`) I would need to utilize the zulip client to get a list
of stream names, for example.
2. I'm not sure how the database responds now that I've changed it. What
would happen if I try and access an old element that does not have the
`topic` field occupied?

Issues:
1. No priorities. At the moment, we can find a pair of recursers because
they have matching goals/topics, but if the goal/topic is rare, then
this recurser would have noone to talk with. It's an all or nothing
approach. I, for example, would like to talk about ML, but if noone
decides to talk about ML on this day, I would not match with anyone. I
would like it so that if I don't match with someone pertaining to a
topic, I can match with anyone else.
2. Only one meeting. At the moment, there is only one meeting per day.
This is related to issue 1 in some sense.