Allow multiple pairings per day on different topics
The motivation here is that talking/pairing with a wide variety of
people is fun and an important aspect of recurse to many recursers. To
maximize fun, we can then give the recursers the options of having more
than one pairing per day on differnt topics.

The original behaviour (one pairing with anyone per day) is still the
default behaviour, so I don't think this breaks anything. It hasn't been
tested yet though.

I believe that giving recursers the option to choose the topic/stream of
the pairing will increase person to person engagement especially within
a stream. Since a lot of topics on zulip seems to be more of Q&A and
status update in spirit, it may be hard to just talk w/o structure or
purpose (like stream of consciousness). I think direct pairing to talk
maybe a nice way to fill in this gap.

Since the pairing bot simply starts the dialog between recursers, it
shouldn't get too spammy (hopefully). And even if it does, one can
simply reduce the frequency of meetings

What:
1. Replace topic (string) attribute/key to streams (map from string to
int), which counts the number of meetings per day per stream/topic
2. Add some string formats to make status output and help message pretty
3. Add input check for streams and counts
4. Modify N^2 pairing algorithm to match by stream. Still N^2 I think or
maybe it's M where M is the total number of allowed pairs